pyproject.toml
[tool.poetry]
name = "tiberium"
version = "0.1.0"
description = ""
authors = ["Aleksey Ryabchikov <cyberxx@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
djangorestframework = "^3.9"
celery = "^4.3.0rc3"
flower = "^0.9.3"
drf-yasg = "^1.14"
psycopg2 = "^2.7"
redis = "^3.2"
faker = "^2.0"
djangorestframework-jwt = "^1.11"

Pillow = "^6.0"
django-cors-headers = "^2.5"
djangorestframework-recaptcha = "^0.2"
django_redis = "^4.10"
hiredis = "^1.0"
gunicorn = "^19.9"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
pytest = "^4.3"
pytest-django = "^3.4"
black = "=19.3b0"
isort = "^4.3"
pylint = "^2.2"
mypy = "^0.720"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
